# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/families": {
      "get": {
        "operationId": "FamiliesController_listFamilies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/families/{id}": {
      "get": {
        "operationId": "FamiliesController_getFamilyDetail",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      },
      "patch": {
        "operationId": "FamiliesController_updateFamily",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/families/{id}/guardians": {
      "post": {
        "operationId": "FamiliesController_addGuardianToFamily",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/families/{id}/students": {
      "post": {
        "operationId": "FamiliesController_addStudentToFamily",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/families/wizard": {
      "post": {
        "operationId": "FamiliesController_createFamilyWizard",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/guardians": {
      "get": {
        "operationId": "FamiliesController_listGuardians",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/guardians/{id}/create-login": {
      "post": {
        "operationId": "FamiliesController_createGuardianLogin",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/guardians/{id}/link-user": {
      "post": {
        "operationId": "FamiliesController_linkGuardianUser",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/students": {
      "get": {
        "operationId": "FamiliesController_listStudents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/students/{id}": {
      "get": {
        "operationId": "FamiliesController_getStudentDetail",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      },
      "patch": {
        "operationId": "FamiliesController_updateStudent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/guardian-student-relationships/{id}": {
      "patch": {
        "operationId": "FamiliesController_updateGuardianStudentRelationship",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
...
  },
  "info": {
    "title": "Eudora API",
    "description": "REST API for the Eudora monorepo scaffold.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ApiPaginationMeta": {
        "type": "object",
        "required": [
          "page",
          "pageSize",
          "totalItems",
          "totalPages",
          "hasNext",
          "hasPrev"
        ],
        "properties": {
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrev": {
            "type": "boolean"
          }
        }
      },
      "ApiMeta": {
        "type": "object",
        "required": [
          "requestId",
          "timestamp",
          "version",
          "path",
          "method"
        ],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "string",
            "enum": [
              "v1"
            ]
          },
          "path": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "pagination": {
            "$ref": "#/components/schemas/ApiPaginationMeta"
          }
        }
      },
      "ApiEnvelope": {
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "data",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetail"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      }
    }
  }
}